Move modules class to separate file modules.py
